+++ public/app.js
@@ -1303,11 +1303,6 @@ __webpack_require__.r(__webpack_exports__);
 //
 //
 //
-//
-//
-//
-//
-//
 /* harmony default export */ __webpack_exports__["default"] = ({
   name: 'MenuGroup',
   props: {
@@ -1344,6 +1339,8 @@ __webpack_require__.r(__webpack_exports__);
 //
 //
 //
+//
+//
 
 /* harmony default export */ __webpack_exports__["default"] = ({
   name: 'MenuLink',
@@ -1356,7 +1353,14 @@ __webpack_require__.r(__webpack_exports__);
   computed: {
     classes: function classes() {
       if (this.$page.app.requestUrl.startsWith(lodash_trimStart__WEBPACK_IMPORTED_MODULE_0___default()(this.item.href, this.$page.app.baseUrl))) {
-        return 'text-gray-800 border-r-8 border-brand-500';
+        return 'text-gray-800  border-r-8 border-brand-500';
+      }
+
+      return 'bg-transparent';
+    },
+    containerClasses: function containerClasses() {
+      if (this.$page.app.requestUrl.startsWith(lodash_trimStart__WEBPACK_IMPORTED_MODULE_0___default()(this.item.href, this.$page.app.baseUrl))) {
+        return 'bg-gray-100 border-t border-b border-gray-400';
       }
 
       return '';
@@ -3042,6 +3046,25 @@ exports = module.exports = __webpack_require__(/*! ../../../../node_modules/css-
 
 // module
 exports.push([module.i, ".form-input.v-select .vs__selected {\n  padding: 2px 4px !important;\n  margin-right: 4 px !important;\n}\n.vs__selected {\n  font-size: 0.75rem !important;\n  font-weight: 600 !important;\n  color: #2d3748 !important;\n  background-color: #f2f5fa !important;\n  border-color:  !important;\n}\n", ""]);
+
+// exports
+
+
+/***/ }),
+
+/***/ "./node_modules/css-loader/index.js?!./node_modules/vue-loader/lib/loaders/stylePostLoader.js!./node_modules/postcss-loader/src/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/Components/MenuGroup.vue?vue&type=style&index=0&id=034c12a5&scoped=true&lang=css&":
+/*!***************************************************************************************************************************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/css-loader??ref--6-1!./node_modules/vue-loader/lib/loaders/stylePostLoader.js!./node_modules/postcss-loader/src??ref--6-2!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/Components/MenuGroup.vue?vue&type=style&index=0&id=034c12a5&scoped=true&lang=css& ***!
+  \***************************************************************************************************************************************************************************************************************************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports, __webpack_require__) {
+
+exports = module.exports = __webpack_require__(/*! ../../../node_modules/css-loader/lib/css-base.js */ "./node_modules/css-loader/lib/css-base.js")(false);
+// imports
+
+
+// module
+exports.push([module.i, ".menu-group__label[data-v-034c12a5]::before {\n  border-top-width: 1px;\n  border-style: dashed;\n  --border-opacity: 1;\n  border-color: #cbd5e0;\n  border-color: rgba(203, 213, 224, var(--border-opacity));\n  content: ' ';\n  display: block;\n  position: absolute;\n  width: 100%;\n}\n", ""]);
 
 // exports
 
@@ -28679,6 +28702,36 @@ if(false) {}
 
 /***/ }),
 
+/***/ "./node_modules/style-loader/index.js!./node_modules/css-loader/index.js?!./node_modules/vue-loader/lib/loaders/stylePostLoader.js!./node_modules/postcss-loader/src/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/Components/MenuGroup.vue?vue&type=style&index=0&id=034c12a5&scoped=true&lang=css&":
+/*!*******************************************************************************************************************************************************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/style-loader!./node_modules/css-loader??ref--6-1!./node_modules/vue-loader/lib/loaders/stylePostLoader.js!./node_modules/postcss-loader/src??ref--6-2!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/Components/MenuGroup.vue?vue&type=style&index=0&id=034c12a5&scoped=true&lang=css& ***!
+  \*******************************************************************************************************************************************************************************************************************************************************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports, __webpack_require__) {
+
+
+var content = __webpack_require__(/*! !../../../node_modules/css-loader??ref--6-1!../../../node_modules/vue-loader/lib/loaders/stylePostLoader.js!../../../node_modules/postcss-loader/src??ref--6-2!../../../node_modules/vue-loader/lib??vue-loader-options!./MenuGroup.vue?vue&type=style&index=0&id=034c12a5&scoped=true&lang=css& */ "./node_modules/css-loader/index.js?!./node_modules/vue-loader/lib/loaders/stylePostLoader.js!./node_modules/postcss-loader/src/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/Components/MenuGroup.vue?vue&type=style&index=0&id=034c12a5&scoped=true&lang=css&");
+
+if(typeof content === 'string') content = [[module.i, content, '']];
+
+var transform;
+var insertInto;
+
+
+
+var options = {"hmr":true}
+
+options.transform = transform
+options.insertInto = undefined;
+
+var update = __webpack_require__(/*! ../../../node_modules/style-loader/lib/addStyles.js */ "./node_modules/style-loader/lib/addStyles.js")(content, options);
+
+if(content.locals) module.exports = content.locals;
+
+if(false) {}
+
+/***/ }),
+
 /***/ "./node_modules/style-loader/index.js!./node_modules/css-loader/index.js?!./node_modules/vue-loader/lib/loaders/stylePostLoader.js!./node_modules/postcss-loader/src/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/Components/Spinner.vue?vue&type=style&index=0&id=aba859c4&scoped=true&lang=css&":
 /*!*****************************************************************************************************************************************************************************************************************************************************************************************************************************!*\
   !*** ./node_modules/style-loader!./node_modules/css-loader??ref--6-1!./node_modules/vue-loader/lib/loaders/stylePostLoader.js!./node_modules/postcss-loader/src??ref--6-2!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/Components/Spinner.vue?vue&type=style&index=0&id=aba859c4&scoped=true&lang=css& ***!
@@ -31613,32 +31666,32 @@ var render = function() {
   var _c = _vm._self._c || _h
   return _c(
     "div",
-    {
-      staticClass:
-        "relative w-full py-3 flex items-center font-semibold text-gray-500 hover:text-gray-700 transition-all ease-in-out duration-300 cursor-pointer opacity-75 hover:opacity-100",
-      attrs: { href: _vm.item.href },
-      on: {
-        click: function() {
-          return _vm.$emit("click", _vm.item)
-        },
-        mouseenter: function(ev) {
-          return _vm.$emit("mouseenter", ev)
-        },
-        mouseleave: function(ev) {
-          return _vm.$emit("mouseleave", ev)
-        }
-      }
-    },
+    { staticClass: "border-b border-gray-400 border-dashed pb-4" },
     [
       _c(
-        "span",
-        { staticClass: "flex text-2xl justify-center w-16 opacity-75" },
-        [_c("f-icon", { attrs: { name: _vm.item.icon } })],
-        1
+        "div",
+        {
+          staticClass:
+            "relative w-full py-3 flex items-center text-gray-600 opacity-100 menu-group__label"
+        },
+        [
+          _c(
+            "span",
+            {
+              staticClass:
+                "flex-shrink bg-gray-300 text-sm font-semibold uppercase tracking-wider z-10 ml-12 px-4"
+            },
+            [_vm._v(_vm._s(_vm.item.label))]
+          )
+        ]
       ),
       _vm._v(" "),
-      _c("span", { staticClass: "text-sm" }, [_vm._v(_vm._s(_vm.item.label))])
-    ]
+      _c("f-navigation-list", {
+        staticClass: "transition-all duration-300",
+        attrs: { items: _vm.item.children }
+      })
+    ],
+    1
   )
 }
 var staticRenderFns = []
@@ -31664,37 +31717,43 @@ var render = function() {
   var _h = _vm.$createElement
   var _c = _vm._self._c || _h
   return _c(
-    "inertia-link",
-    {
-      staticClass:
-        "relative w-full py-3 flex items-center text-gray-600 hover:text-gray-700 transition-all ease-in-out duration-300 opacity-75 hover:opacity-100",
-      class: _vm.classes,
-      attrs: { href: _vm.item.href },
-      on: {
-        click: function() {
-          return _vm.$emit("click", _vm.item)
-        }
-      }
-    },
+    "div",
+    { class: _vm.containerClasses },
     [
-      _vm.item.icon
-        ? _c(
+      _c(
+        "inertia-link",
+        {
+          staticClass:
+            "relative w-full py-3 flex items-center text-gray-600 hover:text-gray-700 transition-all ease-in-out duration-300 opacity-75 hover:opacity-100",
+          class: _vm.classes,
+          attrs: { href: _vm.item.href },
+          on: {
+            click: function() {
+              return _vm.$emit("click", _vm.item)
+            }
+          }
+        },
+        [
+          _c(
             "span",
             { staticClass: "flex text-2xl font-semibold justify-center w-16" },
-            [_c("f-icon", { attrs: { name: _vm.item.icon } })],
+            [
+              _vm.item.icon
+                ? _c("f-icon", { attrs: { name: _vm.item.icon } })
+                : _vm._e()
+            ],
             1
+          ),
+          _vm._v(" "),
+          _c(
+            "span",
+            { staticClass: "flex-1 text-sm font-medium opacity-100" },
+            [_vm._v(_vm._s(_vm.item.label))]
           )
-        : _vm._e(),
-      _vm._v(" "),
-      _c(
-        "span",
-        {
-          staticClass: "text-sm font-medium opacity-100",
-          class: _vm.item.icon ? "" : "pl-8"
-        },
-        [_vm._v(_vm._s(_vm.item.label))]
+        ]
       )
-    ]
+    ],
+    1
   )
 }
 var staticRenderFns = []
@@ -31721,7 +31780,7 @@ var render = function() {
   var _c = _vm._self._c || _h
   return _c(
     "div",
-    { staticClass: "relative my-6" },
+    { staticClass: "relative" },
     [
       _vm._l(_vm.items, function(item, index) {
         return [
@@ -32041,7 +32100,12 @@ var render = function() {
               staticClass:
                 "flex-grow overflow-y-auto overflow-x-hidden custom-scroll"
             },
-            [_c("f-navigation-list", { attrs: { items: this.$page.menu } })],
+            [
+              _c("f-navigation-list", {
+                staticClass: "my-6",
+                attrs: { items: this.$page.menu }
+              })
+            ],
             1
           )
         ]
@@ -33273,6 +33337,7 @@ var render = function() {
                     },
                     [
                       _c("f-navigation-list", {
+                        staticClass: "my-6",
                         attrs: { items: _vm.sidebarMenu }
                       })
                     ],
@@ -35392,7 +35457,9 @@ __webpack_require__.r(__webpack_exports__);
 __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var _MenuGroup_vue_vue_type_template_id_034c12a5_scoped_true___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./MenuGroup.vue?vue&type=template&id=034c12a5&scoped=true& */ "./resources/js/Components/MenuGroup.vue?vue&type=template&id=034c12a5&scoped=true&");
 /* harmony import */ var _MenuGroup_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./MenuGroup.vue?vue&type=script&lang=js& */ "./resources/js/Components/MenuGroup.vue?vue&type=script&lang=js&");
-/* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js");
+/* empty/unused harmony star reexport *//* harmony import */ var _MenuGroup_vue_vue_type_style_index_0_id_034c12a5_scoped_true_lang_css___WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./MenuGroup.vue?vue&type=style&index=0&id=034c12a5&scoped=true&lang=css& */ "./resources/js/Components/MenuGroup.vue?vue&type=style&index=0&id=034c12a5&scoped=true&lang=css&");
+/* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js");
+
 
 
 
@@ -35400,7 +35467,7 @@ __webpack_require__.r(__webpack_exports__);
 
 /* normalize component */
 
-var component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__["default"])(
+var component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_3__["default"])(
   _MenuGroup_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__["default"],
   _MenuGroup_vue_vue_type_template_id_034c12a5_scoped_true___WEBPACK_IMPORTED_MODULE_0__["render"],
   _MenuGroup_vue_vue_type_template_id_034c12a5_scoped_true___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"],
@@ -35429,6 +35496,22 @@ component.options.__file = "resources/js/Components/MenuGroup.vue"
 __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_MenuGroup_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/babel-loader/lib??ref--4-0!../../../node_modules/vue-loader/lib??vue-loader-options!./MenuGroup.vue?vue&type=script&lang=js& */ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/Components/MenuGroup.vue?vue&type=script&lang=js&");
 /* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__["default"] = (_node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_MenuGroup_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__["default"]); 
+
+/***/ }),
+
+/***/ "./resources/js/Components/MenuGroup.vue?vue&type=style&index=0&id=034c12a5&scoped=true&lang=css&":
+/*!********************************************************************************************************!*\
+  !*** ./resources/js/Components/MenuGroup.vue?vue&type=style&index=0&id=034c12a5&scoped=true&lang=css& ***!
+  \********************************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var _node_modules_style_loader_index_js_node_modules_css_loader_index_js_ref_6_1_node_modules_vue_loader_lib_loaders_stylePostLoader_js_node_modules_postcss_loader_src_index_js_ref_6_2_node_modules_vue_loader_lib_index_js_vue_loader_options_MenuGroup_vue_vue_type_style_index_0_id_034c12a5_scoped_true_lang_css___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/style-loader!../../../node_modules/css-loader??ref--6-1!../../../node_modules/vue-loader/lib/loaders/stylePostLoader.js!../../../node_modules/postcss-loader/src??ref--6-2!../../../node_modules/vue-loader/lib??vue-loader-options!./MenuGroup.vue?vue&type=style&index=0&id=034c12a5&scoped=true&lang=css& */ "./node_modules/style-loader/index.js!./node_modules/css-loader/index.js?!./node_modules/vue-loader/lib/loaders/stylePostLoader.js!./node_modules/postcss-loader/src/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/Components/MenuGroup.vue?vue&type=style&index=0&id=034c12a5&scoped=true&lang=css&");
+/* harmony import */ var _node_modules_style_loader_index_js_node_modules_css_loader_index_js_ref_6_1_node_modules_vue_loader_lib_loaders_stylePostLoader_js_node_modules_postcss_loader_src_index_js_ref_6_2_node_modules_vue_loader_lib_index_js_vue_loader_options_MenuGroup_vue_vue_type_style_index_0_id_034c12a5_scoped_true_lang_css___WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_node_modules_style_loader_index_js_node_modules_css_loader_index_js_ref_6_1_node_modules_vue_loader_lib_loaders_stylePostLoader_js_node_modules_postcss_loader_src_index_js_ref_6_2_node_modules_vue_loader_lib_index_js_vue_loader_options_MenuGroup_vue_vue_type_style_index_0_id_034c12a5_scoped_true_lang_css___WEBPACK_IMPORTED_MODULE_0__);
+/* harmony reexport (unknown) */ for(var __WEBPACK_IMPORT_KEY__ in _node_modules_style_loader_index_js_node_modules_css_loader_index_js_ref_6_1_node_modules_vue_loader_lib_loaders_stylePostLoader_js_node_modules_postcss_loader_src_index_js_ref_6_2_node_modules_vue_loader_lib_index_js_vue_loader_options_MenuGroup_vue_vue_type_style_index_0_id_034c12a5_scoped_true_lang_css___WEBPACK_IMPORTED_MODULE_0__) if(__WEBPACK_IMPORT_KEY__ !== 'default') (function(key) { __webpack_require__.d(__webpack_exports__, key, function() { return _node_modules_style_loader_index_js_node_modules_css_loader_index_js_ref_6_1_node_modules_vue_loader_lib_loaders_stylePostLoader_js_node_modules_postcss_loader_src_index_js_ref_6_2_node_modules_vue_loader_lib_index_js_vue_loader_options_MenuGroup_vue_vue_type_style_index_0_id_034c12a5_scoped_true_lang_css___WEBPACK_IMPORTED_MODULE_0__[key]; }) }(__WEBPACK_IMPORT_KEY__));
+ /* harmony default export */ __webpack_exports__["default"] = (_node_modules_style_loader_index_js_node_modules_css_loader_index_js_ref_6_1_node_modules_vue_loader_lib_loaders_stylePostLoader_js_node_modules_postcss_loader_src_index_js_ref_6_2_node_modules_vue_loader_lib_index_js_vue_loader_options_MenuGroup_vue_vue_type_style_index_0_id_034c12a5_scoped_true_lang_css___WEBPACK_IMPORTED_MODULE_0___default.a); 
 
 /***/ }),