+++ src/Support/Menu/Contracts/MenuItem.php
@@ -4,6 +4,8 @@ namespace CbtechLtd\Fastlane\Support\Menu\Contracts;
 
 interface MenuItem
 {
+    public function getGroup(): string;
+
     public function when(\Closure $whenFn): self;
 
     public function build($user): ?array;