+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to `fastlane` will be documented in this file
 
 ## [Unreleased]
 
+### Added
+- Menu group to admin.
+
+### Changed
+- How entry types links on menu are identified.
+
+---
+
 ## 0.1.7 [2020-08-17]
 
 ### Added